third_party/blink/web_tests/fast/forms/date-multiple-fields/date-multiple-fields-non-disabled-subfield.html

<!DOCTYPE html>
<title>Sub-fields in input[type=week] should not be disabled</title>
<link rel='help' href='https://issues.chromium.org/issues/442293382'>
<script src='../../../resources/testharness.js'></script>
<script src='../../../resources/testharnessreport.js'></script>
<script src='../resources/common.js'></script>
<script>
function createDateInput(min, max, value) {
  return createInput('date', min, max, value);
}

test(()=>{
  assert_false(isYearFieldDisabled(createDateInput('', '', '')));
  assert_false(isYearFieldDisabled(createDateInput('2012-12-17', '', '')));
  assert_false(isYearFieldDisabled(createDateInput('', '2012-12-17', '')));
  assert_false(isYearFieldDisabled(createDateInput('', '', '2012-12-17')));
  assert_false(isYearFieldDisabled(createDateInput('2012-01-01', '2012-12-31', '')));
  assert_false(isYearFieldDisabled(createDateInput('2012-01-01', '2013-12-31', '')));
  assert_false(isYearFieldDisabled(createDateInput('2012-01-01', '2012-12-31', '2012-05-30')));
  assert_false(isYearFieldDisabled(createDateInput('2012-01-01', '2012-12-31', '2013-01-01')));
}, 'The year field should be non-disabled');

test(()=>{
  assert_false(isMonthFieldDisabled(createDateInput('', '', '')));
  assert_false(isMonthFieldDisabled(createDateInput('2012-12-17', '', '')));
  assert_false(isMonthFieldDisabled(createDateInput('', '2012-12-17', '')));
  assert_false(isMonthFieldDisabled(createDateInput('', '', '2012-12-17')));
  assert_false(isMonthFieldDisabled(createDateInput('2012-12-01', '2012-12-31', '')));
  assert_false(isMonthFieldDisabled(createDateInput('2012-11-01', '2013-12-31', '')));
  assert_false(isMonthFieldDisabled(createDateInput('2012-12-01', '2013-12-31', '')));
  assert_false(isMonthFieldDisabled(createDateInput('2012-12-01', '2012-12-31', '2012-12-15')));
  assert_false(isMonthFieldDisabled(createDateInput('2012-12-01', '2012-12-31', '2012-11-30')));
}, 'The month field should be non-disabled');

test(()=>{
  assert_false(isDayFieldDisabled(createDateInput('', '', '')));
  assert_false(isDayFieldDisabled(createDateInput('2012-12-17', '', '')));
  assert_false(isDayFieldDisabled(createDateInput('', '2012-12-17', '')));
  assert_false(isDayFieldDisabled(createDateInput('', '', '2012-12-17')));
  assert_false(isDayFieldDisabled(createDateInput('2012-12-17', '2012-12-17', '')));
  assert_false(isDayFieldDisabled(createDateInput('2012-12-17', '2013-12-18', '')));
  assert_false(isDayFieldDisabled(createDateInput('2012-11-17', '2012-12-17', '')));
  assert_false(isDayFieldDisabled(createDateInput('2012-12-17', '2013-12-17', '')));
  assert_false(isDayFieldDisabled(createDateInput('2012-12-17', '2012-12-17', '2012-12-17')));
  assert_false(isDayFieldDisabled(createDateInput('2012-12-17', '2012-12-17', '2012-12-18')));
}, 'The day field should be non-disabled');
</script>
